Modify process_sra_hosted to run locally (#33)
This pull request includes several updates to the GitHub Actions
workflows and configuration files to improve the setup and execution of
the CI pipeline. The changes mainly focus on conditional execution of
steps based on the environment and simplifying the configuration.

### Improvements to GitHub Actions workflows:

*
[`.github/actions/install-dependencies/action.yml`](diffhunk://#diff-6d8cc6283ab3a0a0d3db73624febe8bd91ec9ca7e229c25537ea2f471c2721bfL6-R13):
Added conditions to execute certain steps only when running on
GitHub-hosted runners, such as freeing disk space and setting up conda
and nextflow.
[[1]](diffhunk://#diff-6d8cc6283ab3a0a0d3db73624febe8bd91ec9ca7e229c25537ea2f471c2721bfL6-R13)
[[2]](diffhunk://#diff-6d8cc6283ab3a0a0d3db73624febe8bd91ec9ca7e229c25537ea2f471c2721bfR27)
[[3]](diffhunk://#diff-6d8cc6283ab3a0a0d3db73624febe8bd91ec9ca7e229c25537ea2f471c2721bfR37)

*
[`.github/workflows/process_sra_hosted.yml`](diffhunk://#diff-837d1cb85146d1ae7b898b2d30b3467a55e29f76a7c28742df6e0925aeb75180R21-R22):
Added environment variables `NXF_WORK` and `NXF_OUTPUT` to the workflow.

### Simplification of configuration:

*
[`.github/workflows/process_sra_hosted.yml`](diffhunk://#diff-837d1cb85146d1ae7b898b2d30b3467a55e29f76a7c28742df6e0925aeb75180L32-R59):
Removed container configuration and several setup steps, relying on the
`install-dependencies` action instead. Updated paths to use `${{
github.workspace }}` for consistency.
[[1]](diffhunk://#diff-837d1cb85146d1ae7b898b2d30b3467a55e29f76a7c28742df6e0925aeb75180L32-R59)
[[2]](diffhunk://#diff-837d1cb85146d1ae7b898b2d30b3467a55e29f76a7c28742df6e0925aeb75180L95-R87)

*
[`config/nextflow.config`](diffhunk://#diff-7f94e6861bbc271f99d3db508b5f2cb65ee87c7a726f2d31735d97876853aeecL10-L14):
Removed resource limits for running on GitHub Actions.