Update output path checks in GitHub Actions workflow for metadata files (#37)
This pull request includes changes to the
`.github/workflows/process_sra_hosted.yml` file to update the conditions
for pushing outputs and generating run summaries. The changes ensure
that the workflow checks for the presence of specific metadata files in
the `outputs/metadata` directory instead of using the `NXF_OUTPUT`
environment variable.

Updates to workflow conditions:

*
[`.github/workflows/process_sra_hosted.yml`](diffhunk://#diff-837d1cb85146d1ae7b898b2d30b3467a55e29f76a7c28742df6e0925aeb75180L46-R46):
Updated the condition for the "Push outputs" step to check for the
presence of `.tsv` files in the `outputs/metadata` directory.
*
[`.github/workflows/process_sra_hosted.yml`](diffhunk://#diff-837d1cb85146d1ae7b898b2d30b3467a55e29f76a7c28742df6e0925aeb75180L67-R67):
Updated the condition for the "Generate Run Summary" step to check for
the presence of `.tsv` files in the `outputs/metadata` directory.