Update cron schedule for GenBank mapping workflow (#42)
This pull request includes updates to workflow files to adjust the
schedule and improve the formatting of the GitHub step summary.

Changes to workflow schedules and formatting:

*
[`.github/workflows/map_to_genbank_seqs.yml`](diffhunk://#diff-965236539dbec3ca3d7a3327f8fe0b61e1e5fe052d05d938be942eab07f4b320L5-R5):
Updated the cron schedule from '0 18 * * *' to '15 1 * * *' to change
the timing of the scheduled run.
*
[`.github/workflows/process_sra_hosted.yml`](diffhunk://#diff-837d1cb85146d1ae7b898b2d30b3467a55e29f76a7c28742df6e0925aeb75180R79-R84):
Added blank lines to improve the readability of the GitHub step summary
by separating sections more clearly.