Refactor output handling in GitHub Actions workflow to streamline metadata and file processing (#43)
This pull request includes changes to the
`.github/workflows/process_sra_hosted.yml` file to improve the handling
and processing of output files. The changes focus on configuring Git,
moving and committing output files, and ensuring the workflow runs
smoothly.

Improvements to file handling and Git configuration:

* Added Git configuration steps to set the user email and name for
GitHub Actions.
* Modified the logic to move and commit output files by using loops for
better readability and maintainability. This includes handling `fasta`,
`depth`, and `variants` directories in a loop instead of individually.
* Ensured the `demixed` directory is processed separately and added to
Git if it exists.
* Added comments to clarify the steps for moving updated metadata and
adding files to Git.
* Removed the conditional check for pushing outputs based on the
presence of metadata files.