Add bioproject description in README (#51)
This pull request includes changes to the GitHub workflow and updates to
the `README.md` file to improve the clarity and functionality of the
project documentation and automation processes. The most important
changes include adding a condition to the GitHub workflow to ensure only
non-empty metadata files trigger a push, and reformatting the
`README.md` to include descriptions for each BioProject.

### Improvements to GitHub workflow:

*
[`.github/workflows/process_sra_hosted.yml`](diffhunk://#diff-837d1cb85146d1ae7b898b2d30b3467a55e29f76a7c28742df6e0925aeb75180L73-R73):
Added a condition to the push step to only proceed if the metadata CSV
files are not empty.

### Documentation updates:

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L3-R12):
Reformatted the BioProjects section into a table and added descriptions
for each BioProject to improve readability and provide more context.