Refactor conditional logic for GitHub Actions workflow to simplify push criteria (#52)
This pull request includes a minor change to the
`.github/workflows/process_sra_hosted.yml` file. The change simplifies
the condition for the "Push to GitHub" step by removing the check for
non-empty CSV files in the `outputs/metadata` directory.

*
[`.github/workflows/process_sra_hosted.yml`](diffhunk://#diff-837d1cb85146d1ae7b898b2d30b3467a55e29f76a7c28742df6e0925aeb75180L73-R73):
Simplified the condition for the "Push to GitHub" step by removing the
check for non-empty CSV files.