Update workflows to trigger on completion of 'Process Flu SRA' and move genoflu results to metadata (#56)
This pull request includes updates to the GitHub Actions workflows to
improve the handling of GenBank sequences and processing of directories.
The key changes are:

Workflow updates:

*
[`.github/workflows/map_to_genbank_seqs.yml`](diffhunk://#diff-965236539dbec3ca3d7a3327f8fe0b61e1e5fe052d05d938be942eab07f4b320L4-R7):
Changed the trigger from a scheduled cron job to a `workflow_run` that
triggers upon the completion of the "Process Flu SRA (self-hosted)"
workflow.

Directory processing improvements:

*
[`.github/workflows/process_sra_hosted.yml`](diffhunk://#diff-837d1cb85146d1ae7b898b2d30b3467a55e29f76a7c28742df6e0925aeb75180R63-R67):
Added a step to move the `genoflu_results.tsv` file from the `genoflu`
directory to the `metadata` directory if the `genoflu` directory exists.